+++ taktik/core/clone/selectors/__init__.py
@@ -1,0 +1,5 @@
+"""Clone selector adaptation owners."""
+
+from taktik.core.clone.selectors.patcher import patch_selectors_for_package
+
+__all__ = ["patch_selectors_for_package"]